backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
openai
python-dotenv
python-multipart
